Top regions & countries for volcanic eruptions
Plus an interesting discovery that there are 50+ cases of recent confirmed eruptions (since year 2000) but whose volcano descriptions are not in the base volcano dataset. I suspect that there is an "official" volcano repository that is up to date (because all eruptions bear a volcano_number), but the volcano data in this TidyTuesday set was not up to date.